Move Article routes to `route`
